Refactor dashboard component imports for improved clarity and update cspell configuration to include 'adsense'
