+++ apps/nextjs/src/app/(dashboard)/dashboard/_components/dashboard-content-skeleton.tsx
