+++ cspell.config.yaml
@@ -37,6 +37,7 @@ dictionaries:
 words:
   - adbanner
   - adsbygoogle
+  - adsense
   - adtest
   - alola
   - battlestadium
